Refactor controller and service methods for getting ancestors path
